refactor: split voice-agent helpers into modular services
