+++ src/lib/agents/realtime/voice-agent/tool-context.ts
@@ -1,0 +1,3 @@
+import { jsonObjectSchema } from '@/lib/utils/json-schema';
+
+export const createToolParametersSchema = () => jsonObjectSchema.default({});
